Add patched version of nvflash v1.13.87205 for Linux (32-bit)
This has a single byte patch to the tegra30 miniloader embedded inside
the nvflash binary. The patch overrides the function that checks
various secure-related fuses on the SoC. More specifically it checks
FUSE_FA, FUSE_SECURITY_MODE, and FUSE_PRODUCTION_MODE.

Without this patch it is not possible for nvflash to successfully upload
and execute its embedded miniloader, which is required for almost all of
the functionality that nvflash provides.