Add tegra30 version of ipatch_rcm_payload.c
This mostly does the same thing as the unmodified ipatch_rcm_payload.c,
except offsets and addresses have been tailored to tegra30.

Both the Ouya and Nexus 7 are supported by this; to get different UART
output, just comment/uncomment the UARTA / UARTD defines. That will also
handle the pinmux and so forth.

An additional define, ODM_DEBRICK, was added out of necessity after
bricking a bricked Nexus 7 (double bricked :)). Flashing a bad BCT (and
accompanying odmdata) can cause the tablet to become unresponsive
outside of RCM mode. It makes no permanent changes to the board, but
intentionally breaks initialization of the MMC controller so that the
problematic BCT isn't loaded, allowing nvflash to work when it would
otherwise be unresponsive.

Moving the UART initialization and print into a single function was done
to allow the function to be more easily patched into the Nexus 7
bootloader while working to debrick the tablet. See elsewhere in
this repository for a pre-patched copy of that bootloader.