Add modified N7 bootloader with UART output force-enabled
While working towards debricking my N7, I was mostly flying blind. The
bootloader had a function which would attempt to send a lot of useful
debug messages via ARM semihosting (but not over the UART I'd tapped
into on the motherboard).

To get some visibility into what was happening, this was patched to
change the function pointer of the original semihost debug function to a
function I hex-edited in. On the first call, the function initializes
the UART hardware, and then sends all of those strings to the UART-A.

The function was placed into an area of memory occupied by the text
section (strings) of the Trusted Foundations firmware. Baud is
*approximately* 115200 (some transmit errors are expected).

The source for the function can be found at bootloader/uart_print.c,
which was based on the UART initialization from ktemkin's
ipatch_rcm_sample.c.