Add Nexus 7 16GB flash.cfg for use with nvflash
This was originally taken from Partitionator's (lol...) post at
https://forum.xda-developers.com/showpost.php?p=48507424&postcount=16.

Note that although this is for the Nexus 7, it makes nvflash happy
enough when launching nv3pserver on the Ouya. There is already a
partition table on the device (even on a bricked one, most likely), so
it doesn't appear to be used for anything except satisfying nvflash's
required arguments.