Add precompiled uart_payload.bin for Ouya
This is compiled with '#define UARTD' so the UART and pinmux are
tailored to the Ouya. The offsets are the same for the Nexus 7, as they
share the same bootrom. It might work OK on the Nexus 7, or it might
hang due to the different pins used for UART. I intend to eventually add
a precompiled version for Nexus 7.