Add Makefile and linker script for uart_payload.c
This produces byte-for-byte identical files to the current .bin files
when built with gcc version 9.3.0 (Ubuntu 9.3.0-17ubuntu1~20.04).

The default '#define UARTD' has been removed from uart_payload.c and it
now gives an error if neither UARTA or UARTD or defined.

Linker script is based on ../fusee-launcher/dump-sbk-via-usb.lds.

Fixes #1.